Renaming: blindai.client => blindai, BlindAiConnection => onnection
