+++ client/tests/test_requests.py
@@ -2,7 +2,7 @@ import os
 from typing import Iterator
 import unittest
 from unittest.mock import MagicMock, Mock, patch
-import blindai.client
+import blindai
 from unittest.mock import *
 from datetime import datetime, timedelta
 from typing import Any, Dict, List, Optional, Tuple, Union
@@ -20,7 +20,7 @@ from blindai.pb.securedexchange_pb2 import (
     TensorInfo,
 )
 
-from blindai.client import (
+from blindai import (
     ModelDatumType,
     PredictResponse,
     UploadModelResponse,
@@ -51,8 +51,8 @@ class TensorInfoMatcher:
 
 
 class TestRequest(unittest.TestCase):
-    @patch("blindai.client.AttestationStub")
-    @patch("blindai.client.secure_channel")
+    @patch("blindai.AttestationStub")
+    @patch("blindai.secure_channel")
     @patch("time.time", mock_time)
     def test_connect(self, secure_channel: MagicMock, AttestationStub: MagicMock):
         res = UploadModelResponse()
@@ -60,15 +60,15 @@ class TestRequest(unittest.TestCase):
 
         attestation = res.attestation
         AttestationStub().GetSgxQuoteWithCollateral = Mock(return_value=attestation)
-        blindai.client.connect(
+        blindai.connect(
             "localhost",
             policy=os.path.join(os.path.dirname(__file__), "policy.toml"),
             certificate=os.path.join(os.path.dirname(__file__), "host_server.pem"),
         )
 
-    @patch("blindai.client.ExchangeStub")
-    @patch("blindai.client.AttestationStub")
-    @patch("blindai.client.secure_channel")
+    @patch("blindai.ExchangeStub")
+    @patch("blindai.AttestationStub")
+    @patch("blindai.secure_channel")
     @patch("time.time", mock_time)
     def test_upload_model(
         self,
@@ -88,7 +88,7 @@ class TestRequest(unittest.TestCase):
         # get_server_certificate.return_value = attestation.
         attestation = res.attestation
         AttestationStub().GetSgxQuoteWithCollateral = Mock(return_value=attestation)
-        client = blindai.client.connect(
+        client = blindai.connect(
             "localhost",
             policy=os.path.join(os.path.dirname(__file__), "policy.toml"),
             certificate=os.path.join(os.path.dirname(__file__), "host_server.pem"),
@@ -150,9 +150,9 @@ class TestRequest(unittest.TestCase):
         send_model_util(sign=False)
         send_model_util(sign=True)
 
-    @patch("blindai.client.ExchangeStub")
-    @patch("blindai.client.AttestationStub")
-    @patch("blindai.client.secure_channel")
+    @patch("blindai.ExchangeStub")
+    @patch("blindai.AttestationStub")
+    @patch("blindai.secure_channel")
     @patch("time.time", mock_time)
     def test_run_model(
         self,
@@ -171,7 +171,7 @@ class TestRequest(unittest.TestCase):
 
         attestation = res.attestation
         AttestationStub().GetSgxQuoteWithCollateral = Mock(return_value=attestation)
-        client = blindai.client.connect(
+        client = blindai.connect(
             "localhost",
             policy=os.path.join(os.path.dirname(__file__), "policy.toml"),
             certificate=os.path.join(os.path.dirname(__file__), "host_server.pem"),
@@ -239,9 +239,9 @@ class TestRequest(unittest.TestCase):
         client.close()
         self.assertTrue(client.closed)
 
-    @patch("blindai.client.ExchangeStub")
-    @patch("blindai.client.AttestationStub")
-    @patch("blindai.client.secure_channel")
+    @patch("blindai.ExchangeStub")
+    @patch("blindai.AttestationStub")
+    @patch("blindai.secure_channel")
     @patch("ssl.get_server_certificate")
     @patch("time.time", mock_time)
     def test_run_model_simulation(
@@ -270,7 +270,7 @@ class TestRequest(unittest.TestCase):
         )
         AttestationStub().GetCertificate = Mock(return_value=response)
 
-        client = blindai.client.connect("localhost", simulation=True)
+        client = blindai.connect("localhost", simulation=True)
 
         # predict
